# WP8 benchmark lookup optimisation