--- a/lsm-tree.cabal
+++ b/lsm-tree.cabal
@@ -505,12 +505,10 @@ benchmark lsm-tree-bench-wp8
   build-depends:
     , async
     , base
-    , binary
     , bytestring
     , clock
     , containers
     , contra-tracer
-    , cryptohash-sha256
     , deepseq
     , fs-api
     , lsm-tree
@@ -519,6 +517,8 @@ benchmark lsm-tree-bench-wp8
     , lsm-tree:mcg
     , optparse-applicative
     , pretty-show
+    , primitive
+    , random
     , vector
 
   ghc-options:    -rtsopts -with-rtsopts=-T -threaded